+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # NEXT RELEASE
+* Rename `epstein_search` to `epstein_grep`
 * `--email-body` argument
 * `EmailHeader.reply_to` field